api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
redis>=5.0
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
orjson>=3.9
sse-starlette>=2.1
fpdf2>=2.7
pillow>=10.0
